src/backend/app/requirements.txt
fastapi
alembic
emails
uvicorn
sqlmodel
typer
dynaconf
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
pydantic
pydantic[email]
openai